Sync changes made in casacore and aocommon
